+++ src/client/modules/users/authentication/login/login.page.js
@@ -25,9 +25,9 @@ export default class LoginPage extends React.Component {
         return (
             <div id='auth-container' className='login' style={{marginTop: 75}}>
 
-                  <div className="col-sm-7 col-sm-offset-3 col-xs-10 col-xs-offset-1" >
-                      <LoginForm ref={(c) => this.loginForm = c} onSuccess = {::this.back}  />
-                  </div>
+                <div className="col-sm-7 col-sm-offset-3 col-xs-10 col-xs-offset-1" >
+                    <LoginForm ref={(c) => this.loginForm = c} onSuccess = {::this.back}  />
+                </div>
 
             </div>
         );