+++ CHANGELOG.md
@@ -40,7 +40,7 @@ All notable changes to this project will be documented in this file.
 - Add a sample GraphQL endpoint [localhost:3000/graphql](https://localhost:3000/graphql)
 - Change the default Node.js server port from `5000` to `3000`
 - Add a JWT-based authentication cookies (see `src/server.js`)
-- Add a reference implementation of Facebook authentication strategy (`src/core/passport.js`)
+- Add a reference implementation of Facebook authentication strategy (`src/core/Passport.js`)
 - Add a sample database client utility for PostgreSQL (`src/core/db.js`)
 - Optimize the `tools/start.js` script that launches dev server with Browsersync and HMR
 - Replace Superagent with WHATWG Fetch library